requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn
streamlit
pyyaml
shap
mlflow
sentence-transformers
torch
